pyproject.toml
[tool.poetry]
name = "honeycomb-opentelemetry"
version = "0.1.0"
description = "Honeycomb OpenTelemetry Distro for Python"
authors = ["Honeycomb <support@honeycomb.io>"]
readme = "README.md"
packages = [{include = "honeycomb" }]

[tool.poetry.dependencies]
python = "^3.7, >= 3.7.2"
opentelemetry-api = "^1.14.0"
opentelemetry-sdk = "^1.14.0"
opentelemetry-exporter-otlp = "^1.14.0"
opentelemetry-instrumentation = "^0.35b0"

[tool.poetry.group.dev.dependencies]
coverage = "^6.5.0"
pytest = "^7.2.0"
pylint = "^2.15.7"
pycodestyle = "^2.10.0"
importlib-metadata = { version = ">=0.12", python = "<3.8" }
requests-mock = "^1.10.0"

[tool.poetry.plugins."opentelemetry_distro"]
distro = "honeycomb.opentelemetry.distro:HoneycombDistro"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
